[fix] 나중에 마저 고치자
오류가 많이 생김

+++ Assets/Scripts/UI/Battle/Card.cs
@@ -45,7 +45,7 @@ public class Card : NetworkBehaviour
         //NewGameManager에 번호를 localPlayer의 Netid와 함께 넘기기(이건 NewGameManager에서 해줌)
         int cardNum = int.Parse(Text_Number.text);
         //NetworkClient.localPlayer
-        int netId = (int)NetworkClient.localPlayer.netId;
+        uint netId = NetworkClient.localPlayer.netId;
         NewGameManager.Instance.CmdAddSubmittedCard_OnClick_Card(netId, cardNum);
 
         //그리고 선택창 닫기(이건 모든 클라에서 이루어져야해)
